+++ src/Makefile
@@ -1,0 +1,8 @@
+# Writen by: Oscar Bergström
+# https://github.com/OSCARJFB
+
+main: main.c
+	cc main.c validateBrackets.c validateComments.c -o main.o
+
+clean:
+	rm -rf *.o
