+++ src/main.h
@@ -1,0 +1,9 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+#define BUFFER_SIZE 1000
+#define MSG_SIZE 100
+
+void getTextInput(char buffer[BUFFER_SIZE]); 
+
+#endif // MAIN_H
